update connector :
1) Multiple BOM 
Logic : When an Item is produced by more than 1 BOM, frePPLe will create
Alternate Operation with corresponding sub-operations.  
2) Supplier and Item Supplier 
Logic : The supplier from Odoo are now imported in frePPLe. For each
product we can specify a supplier which will be imported as an item
supplier in frePPLe. 
3) SO Modification 
Logic : The SO from frePPLe are no more linked to the SO lines in Odoo
instead they are linked to DO lines. It allows to handle partial
delivery in frePPLe. Quantity already delivered is represented as a
closed demand.